+++ src/renderer/app.vue
@@ -145,6 +145,7 @@
       dispatch('LINTEN_FOR_PRINT_SERVICE_CLEARUP')
       dispatch('LINTEN_FOR_EXPORT_SUCCESS')
       dispatch('LISTEN_FOR_SET_TEXT_DIRECTION')
+      dispatch('LISTEN_FOR_FILE_CHANGE')
       dispatch('LISTEN_FOR_TEXT_DIRECTION_MENU')
       // module: notification
       dispatch('LISTEN_FOR_NOTIFICATION')