+++ .github/CHANGELOG.md
@@ -20,6 +20,7 @@ This update **fixes a XSS security vulnerability** when exporting a document.
 - Clicking a link should open it in the browser (#425)
 - Support maxOS `dark mode`, when you change `mode dark or light` in system, Mark Text will change its theme.
 - Add two new themes  Ulysses Light and Graphite Light theme.
+- Watch file changed in tabs and show a notice(autoSave is `false`) or update the file(autoSave is `true`)
 
 **:butterfly:Optimization**